first release v1.0.0

+++ push.sh
@@ -1,17 +1,26 @@
 #!/usr/bin/env bash
-
+  
 pid=$(pidof ssh-agent)
+remote_url=$(git remote -v)
 
-if [[ ! -z  $pid ]]; then
-  # else run this block of code
-  echo "stopping ssh-agent service with pidof: $(pidof ssh-agent)"
-  kill -9 $(pidof ssh-agent)
-  sleep 2
-  echo "starting ssh-agent"
-  eval `ssh-agent`
-  sleep 2
-  echo "setting up ssh-key"
-  ssh-add ~/path/to/ssh/kay
-  sleep 2
-  git commit $1
+if [[ -z $remote_url ]]; then
+  echo "unable to push changes! URL to repository not set"
+  exit 1
+else
+  echo "pushing commit to remote repositry"
+   
+  if [[ ! -z  $pid ]]; then
+    # else run this block of code
+    echo "stopping ssh-agent service with pidof: $(pidof ssh-agent)"
+    kill -9 $(pidof ssh-agent)
+    sleep 2
+    echo "starting ssh-agent"
+    eval `ssh-agent`
+    sleep 2
+    echo "setting up ssh-key"
+    ssh-add ~/path/to/ssh/kay
+    sleep 2
+    git commit $1
+  fi
 fi
+
